fix(sonic): increase the delay-timer instead of disabling it
It should fix an issue in FRR8.2.2 (SONiC 202205).

Before this commit the timer was disabled completely.

We highly suspect that disabling and re-enabling the event driven
changes blocks the updates indefinitely when a prefix-list changes.

When putting 300s instead, once this timer expires, the changes are well
applied. The default 5s are not respected immediately.

Did not have the issue with FRR7.2.1 (SONiC 201911).